spotless

+++ hedera-node/src/main/java/com/hedera/services/state/expiry/removal/NftReturnOutcome.java
@@ -18,4 +18,4 @@ package com.hedera.services.state.expiry.removal;
 import com.hedera.services.utils.EntityNumPair;
 
 public record NftReturnOutcome(
-        NonFungibleTreasuryReturns nftReturns, EntityNumPair newRoot, int remainingNfts) {}
+        NonFungibleTreasuryReturns nftReturns, EntityNumPair newRoot, long remainingNfts) {}
